+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 - Updated Privacy Policy and Terms of Service
 - Reworked collection page UI
 - Empty text symbols now display a placeholder question mark
+- Puzzle cards now show an indication of how the rated difficulty compares to design difficulty
+- The rated difficulty is now rounded up instead of down
 - Fixed errors when switching between puzzle and collection search with filters enabled
 - Fixed the puzzle editor sometimes not loading the newest version of an online puzzle
 - Puzzle completion SFX no longer plays in the editor